+++ src/main/java/it/academy/cursebackspring/services/impl/ProductServiceImpl.java
@@ -2,7 +2,8 @@ package it.academy.cursebackspring.services.impl;
 
 
 import it.academy.cursebackspring.dto.request.CreateProductDTO;
-import it.academy.cursebackspring.dto.request.GetProductPage;
+import it.academy.cursebackspring.dto.request.GetProductPageDTO;
+import it.academy.cursebackspring.dto.request.ParamFilterDTO;
 import it.academy.cursebackspring.dto.request.UpdateProductDTO;
 import it.academy.cursebackspring.dto.response.ProductDTO;
 import it.academy.cursebackspring.dto.response.ProductsDTO;
@@ -15,6 +16,7 @@ import it.academy.cursebackspring.models.Product;
 import it.academy.cursebackspring.repositories.CategoryRepos;
 import it.academy.cursebackspring.repositories.OrderItemRepos;
 import it.academy.cursebackspring.repositories.ProductRepos;
+import it.academy.cursebackspring.repositories.specifications.ProductSpecification;
 import it.academy.cursebackspring.services.ProductService;
 import it.academy.cursebackspring.utilities.Constants;
 import jakarta.validation.Valid;
@@ -23,18 +25,21 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 
-import java.util.Map;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
-@Transactional
+@Transactional(propagation = Propagation.REQUIRED)
 @RequiredArgsConstructor
-@Validated
 public class ProductServiceImpl implements ProductService {
 
     private final CategoryRepos categoryRepos;
@@ -46,7 +51,8 @@ public class ProductServiceImpl implements ProductService {
      *
      * @param createProductDTO important data for method execution
      */
-    public void addProduct(@Valid CreateProductDTO createProductDTO) {
+    @Override
+    public void addProduct(CreateProductDTO createProductDTO) {
         Category category = categoryRepos.findById(createProductDTO.getCategoryId())
                 .orElseThrow(CategoryNotFoundException::new);
         Product product = ProductMapper.INSTANCE.toEntityFromCreateDTO(createProductDTO);
@@ -61,13 +67,14 @@ public class ProductServiceImpl implements ProductService {
      *
      * @param updateProdDTO important data for method execution
      */
-    public void updateProduct(@Valid UpdateProductDTO updateProdDTO) {
-        Optional.of(categoryRepos.existsById(updateProdDTO.getId()))
-                .filter(p -> p)
-                .orElseThrow(CategoryNotFoundException::new);
-        Optional.of(productRepos.existsById(updateProdDTO.getId()))
-                .filter(p -> p)
-                .orElseThrow(ProductNotFoundException::new);
+    @Override
+    public void updateProduct(UpdateProductDTO updateProdDTO) {
+        if (!categoryRepos.existsById(updateProdDTO.getCategoryId())) {
+            throw new CategoryNotFoundException();
+        }
+        if (!productRepos.existsById(updateProdDTO.getId())) {
+            throw new ProductNotFoundException();
+        }
         Product product = ProductMapper.INSTANCE.toEntityFromUpdateDTO(updateProdDTO);
         productRepos.save(product);
     }
@@ -77,10 +84,11 @@ public class ProductServiceImpl implements ProductService {
      *
      * @param id product id
      */
-    public void deleteProduct(@Valid @Min(value = 1, message = Constants.PRODUCT_ID_VALIDATION_EXCEPTION) Long id) {
-        Optional.of(orderItemRepos.existsByOrderItemPK_ProductId(id))
-                .filter(p -> !p)
-                .orElseThrow(ProductUsedInOrdersException::new);
+    @Override
+    public void deleteProduct(Long id) {
+        if (orderItemRepos.existsByOrderItemPK_ProductId(id)) {
+            throw new ProductUsedInOrdersException();
+        }
         productRepos.deleteById(id);
     }
 
@@ -90,32 +98,54 @@ public class ProductServiceImpl implements ProductService {
      * @param id product id
      * @return dto with product info
      */
+    @Override
     @Transactional(readOnly = true)
-    public ProductDTO getProductById(@Valid @Min(value = 1, message = Constants.PRODUCT_ID_VALIDATION_EXCEPTION) Long id) {
+    public ProductDTO getProductById(Long id) {
         Product product = productRepos.findById(id)
                 .orElseThrow(ProductNotFoundException::new);
         return ProductMapper.INSTANCE.toDTOFromEntity(product);
     }
 
     /**
-     * Get the page of products, using filter by category if category id is not null,
-     * method return a page of products filtered on category id.
+     * Get the page of products, using filter by category or name,
+     * method return a filtered page of products.
      *
      * @param dto important data for method execution
      * @return dto which contains list of products and total count of products.
      */
+    @Override
     @Transactional(readOnly = true)
-    public ProductsDTO getAllExistProducts(@Valid GetProductPage dto) {
+    public ProductsDTO getAllExistProducts(GetProductPageDTO dto) {
         PageRequest pageRequest = PageRequest.of(dto.getPageNum(), dto.getCountPerPage());
-        Page<Product> products = (dto.getCategoryId() != null) ?
-                productRepos.findAllByCategoryId(pageRequest, dto.getCategoryId()) :
-                productRepos.findAll(pageRequest);
+        Page<Product> products;
+        List<ParamFilterDTO> filters = new ArrayList<>();
+        if (dto.getNameFilter() != null) {
+            filters.add(toParamFilter(dto.getNameFilter()));
+        }
+        if (dto.getCategoryFilter() != null) {
+            filters.add(toParamFilter(dto.getCategoryFilter()));
+        }
+
+        if (filters.size() > 0) {
+            Specification<Product> spec = new ProductSpecification(filters.get(0));
+            filters.remove(0);
+            for (ParamFilterDTO paramFilterDTO : filters){
+                spec = spec.and(new ProductSpecification(paramFilterDTO));
+            }
+            products = productRepos.findAll(spec, pageRequest);
+        } else {
+            products = productRepos.findAll(pageRequest);
+        }
         return ProductMapper.INSTANCE.toDTOFromEntityList(products.stream().toList(), products.getTotalElements());
     }
 
-
-    public ProductsDTO getAllExistProductByFilterParam(@Valid Integer pageNum,
-                                                       @Valid Integer countPerPage, @Valid Map<String, Object> paramMap) {
-        return null;
+    private ParamFilterDTO toParamFilter(String filterStr) {
+        List<String> operations = List.of(":", "=", ">", "<");
+        String curOperation = operations.stream()
+                .filter(filterStr::contains)
+                .collect(Collectors.joining());
+        String key = filterStr.substring(0, filterStr.indexOf(curOperation));
+        String value = filterStr.substring(filterStr.indexOf(curOperation)+1);
+        return new ParamFilterDTO(curOperation, key, value);
     }
 }
